fix: a compromise!

+++ craft_store/models/track_model.py
@@ -26,7 +26,7 @@ from craft_store.models._base_model import MarshableModel
 class TrackModel(MarshableModel):
     """A track that a package can be published on."""
 
-    automatic_phasing_percentage: int | None = None
+    automatic_phasing_percentage: float | None = None
     created_at: Annotated[  # Prevents pydantic from setting UTC as "...Z"
         datetime,
         pydantic.WrapSerializer(
